[Hexagon] Enable soft bf16 in hexagon
This patch adds:
1. Support to recognize bf16 type in the frontend and isel/abi support
for scalar bf16 programs
Limitations:  fp_to_bf16 is being generated with a tablegen pattern
instead of lowering via expansion. This is because we do not have support for
fcanonincalize instruction which should prevent an SNaN being converted
to an infinity due to truncation.

2. Vector codegen support for bf16

Patch By: Fateme Hosseini

Co-authored-by: Muntasir Mallick <mallick@qti.qualcomm.com>
Co-authored-by: Kaushik Kulkarni <quic_kauskulk@quicinc.com>
Change-Id: I767145458dafcaf7691eb9ab4e03d33e5fd03a6a